add mod_expr to the windows build.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@6190 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/applications/mod_expr/exprpars.c
@@ -97,7 +97,7 @@ int exprStringToTokenList(exprObj *obj, char *expr, exprToken **tokens, int *cou
 
 
     /* Check string length */
-    len = strlen(expr);
+    len = (int)strlen(expr);
     if(len == 0)
         return EXPR_ERROR_EMPTYEXPR;
